Fill in my Resume/CV
Here I provide the the minimal bullets for my work experience and education.

I also added my detailed CV in latex form that I copied from my old GitHub
repo for my old format with static (MS Word created and edited) HTML files.
Because of the way to deploy the site with GitHub, I had to actually commit
the generated cv/bartlett_roscoe_a_cv.pdf file.  I wouild not ordinarily do
this but I need to do this in this case.  I can't imagine that GitHub pages
with Jekyll will be able to generate a latex file for me.